refactor: shared risksRegister read model (dedup RiskRegisterTable/useProjectHub)
PR #466 added byte-faithful copies of RiskRegisterTable's module-private
risk-register helpers into useProjectHub to feed the Overview «Риски —
топ-3» card. Editing one copy silently desynced the top-3 from the
register table's row order (#467).

Extract the read model into src/utils/risksRegister.ts (RISKS_PATH,
RisksFile, enum tables, SEVERITY_RANK, coerceEnum, isIsoCalendarDate,
asString, normaliseRisk, parseRisksFile, sortBySeverityDesc). Both
consumers import from it; local copies deleted. Canonical impl is
RiskRegisterTable's original (the register row order is the source of
truth the Overview must mirror) — behaviour identical, net -231 lines.
